+++ server/lessons/PpcVdfjdzo7TfWBL/evts.js
@@ -27,11 +27,20 @@ return [
             new CP_LineBelow(1),
             new CP_Text("font-style:italic;\nfont-weight:bold;")
         ],"We can also make our text italic and bold by using the font-style and font-weight properties."),
-        new DoRefreshTask("If you haven't already refresh to see all the changes we've made.")
+        new DoRefreshTask("If you haven't already, refresh to see all the changes we've made."),
+        new AddCode([
+            // new CP_MoveBy
+        ],"You can also quickly comment out or disable lines for testing by pressing h[Ctrl/Cmd + /].")
     ]),
     new LE_AddGBubble([
-        "So, now we have all our text customized, what if we wanted to change just h[i[one]] of them?\ni[For example, what if we wanted to highlight our 3rd paragraph element?]"
-    ],BubbleLoc.global,[
-        
+        "So, now we have all our text customized, what if we wanted to change just h[i[one]] of them?\ni[For example, what if we wanted to change the highlight color of our 3rd paragraph element to green?]"
+    ]),
+    new LE_AddGBubble([
+        "Well, in order to do this we can put an id on our element so we can target that one specifically from CSS.",
+        new SwitchFileTask("index.html"),
+        new AddIgnoreCode([
+            new CP_MoveBy(8,9),
+            new CP_Text("(test)")
+        ])
     ])
 ]